move clock
